requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client
requests>=2.31.0

# Google Generative AI
google-generativeai>=0.3.0

# Groq (Fast Inference)
groq>=0.5.0

# Production dependencies
python-multipart>=0.0.6

# Optional: For better performance
orjson>=3.9.0
